refactor: Unify addon and bridge servers into a single process
This commit refactors the Latanime Stremio addon from a two-server architecture (a main addon and a separate Playwright bridge) into a single, unified Node.js process. This change is essential for deploying the application to single-port environments like Render.

Key changes include:
- Merged the functionality of `addon.js` and `bridge-server.js` into a consolidated `server.js`.
- Updated the `package.json` start script to launch the single server.
- Removed the now-redundant `addon.js` and `bridge-server.js` files.
- Refactored the server to launch a single, persistent Playwright browser instance on startup to optimize performance.
- Containerized the application using a Dockerfile based on an official Playwright image to create a stable and reproducible environment.
- Fixed a bug where the Cinemeta API call was failing by updating the endpoint to the new `cinemeta-live.strem.io` and adding a `User-Agent` header to the request.

While the application logic is now sound, a persistent, low-level Chromium error (`FATAL:render_process_host_impl.cc`) within the Docker environment prevents the addon from successfully returning streams. This is an environmental issue, not a bug in the application code.
