--- addon.js
@@ -1,8 +1,0 @@
-const { addonBuilder } = require("stremio-addon-sdk");
-const manifest = require("./lib/manifest");
-const { defineHandlers } = require("./lib/handlers");
-
-const builder = new addonBuilder(manifest);
-defineHandlers(builder);
-
-module.exports = builder.getInterface();
