+++ lib/scraping.js
@@ -9,21 +9,26 @@ const SB_API_KEY = "8MI4VBHDP2PUDO8WU39BC7P2LDSSJ69KX5L5ORQQS0YGKBM73JP46FSNT2DE
 const sbClient = new ScrapingBeeClient(SB_API_KEY);
 const cache = new NodeCache({ stdTTL: 86400 });
 
-async function fetchWithScrapingBee(url, render_js = false) {
-    const cacheKey = `${url}:${render_js}`;
+async function fetchWithScrapingBee(url, render_js = false, json_response = false) {
+    const cacheKey = `${url}:${render_js}:${json_response}`;
     const cached = cache.get(cacheKey);
     if (cached) {
         console.log(`Cache hit for ${cacheKey}`);
         return cached;
     }
 
-    console.log(`Fetching ${url} via ScrapingBee (JS: ${render_js})`);
+    console.log(`Fetching ${url} via ScrapingBee (JS: ${render_js}, JSON: ${json_response})`);
     try {
+        const params = {
+            render_js: render_js.toString(),
+        };
+        if (json_response) {
+            params.json_response = 'true';
+        }
+
         const response = await sbClient.get({
             url: url,
-            params: {
-                render_js: render_js.toString(),
-            },
+            params: params,
         });
 
         if (response.status !== 200) {
