+++ package.json
@@ -4,16 +4,15 @@
   "description": "Provides streams from latanime.org",
   "main": "server.js",
   "scripts": {
-    "start": "concurrently \"node server.js\" \"node bridge-server.js\""
+    "start": "node server.js"
   },
   "dependencies": {
     "axios": "^1.7.2",
     "cheerio": "^1.0.0-rc.12",
-    "concurrently": "^8.2.2",
     "cors": "^2.8.5",
     "express": "^4.19.2",
     "node-cache": "^5.1.2",
-    "playwright": "^1.45.0",
+    "playwright": "1.44.0",
     "scrapingbee": "^1.7.6",
     "stremio-addon-sdk": "^1.6.10"
   }
